requirements.txt
feedparser
requests
beautifulsoup4
trafilatura
lark-oapi
openai
python-dotenv
